--- API/search/basicSearch/index.js
@@ -1,6 +1,0 @@
-const get = require("./get");
-
-module.exports = {
-  getProfileByName: get.getProfileByName,
-  getFuzzySearch: get.getFuzzySearch
-};
