+++ API/search/util/index.js
@@ -1,3 +1,13 @@
-const { fuzzySearch } = require("./fuzzySearch");
+const getAllProfiles = require("./getAllProfiles");
+const fuzzySearch = require("./fuzzySearch");
+const nameSearch = require("./nameSearch");
+const locationSearch = require("./locationSearch");
+const skillSearch = require("./skillSearch");
 
-module.exports = { fuzzySearch };
+module.exports = {
+  getAllProfiles,
+  fuzzySearch,
+  nameSearch,
+  locationSearch,
+  skillSearch
+};
