+++ server.js
@@ -110,11 +110,7 @@ router.use("/option", options);
 router.get("/profGen/:id", profileGeneration.getGedsAssist);
 
 // Search routes
-router.get(
-  "/search/basicSearch/:searchValue",
-  search.basicSearch.getProfileByName
-);
-router.get("/search/fuzzySearch/", search.basicSearch.getFuzzySearch);
+router.get("/search/fuzzySearch/", search);
 
 // REGISTER OUR ROUTES ===============================================
 // Note: All of our routes will be prefixed with /api
